# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MovieFlix Api",
        "description": "Api de gerenciamento dos filmes do MovieFlix",
        "version": "1.0.0"
    },
    "paths": {
        "/movies": {
            "get": {
                "tags": ["Movie"],
                "summary": "Obter todos os filmes",
                "responses": {
                    "200": {
                        "description": "Retorna uma lista de filmes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "put": {
                "tags": ["Movie"],
                "summary": "Atualizar um filme",
                "parameters": {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID do filme a ser atualizado",
                    "schema": {
                        "type": "integer"
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Filme atualizado com sucesso"
                    },
                    "404": {
                        "description": "Registro de filme não encontrado"
                    },
                    "500": {
                        "description": "Falha ao atualizar registro do filme"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "ID do filme"
                    },
                    "title": {
                        "type": "string",
                        "description": "Nome do filme"
                    },
                    "genres_id": {
                        "type": "integer",
                        "description": "Gênero do filme"
                    },
                    "genres": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description":"ID do gênero"
                            },
                            "genre_name": {
                                "type": "string",
                                "description": "Nome do gênero"
                            }
                        }
                    },
                    "languages_id": {
                        "type": "integer",
                        "description": "Linguagem do filme"
                    },
                    "languages": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description":"ID da linguagem"
                            },
                            "language_name": {
                                "type": "string",
                                "description": "Nome da linguagem"
                            }
                        }
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    }
                }
            },
            "MovieInput": {
                "type": "object",
                "required": [
                    "title",
                    "genre_id",
                    "languages_id",
                    "release_date"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Título do filme"
                    },
                    "genres_id": {
                        "type": "integer",
                        "description": "ID do gênero do filme"
                    },
                    "languages_id": {
                        "type": "integer",
                        "description": "ID da língua do filme"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    }
                }
            }
        }
    }
}